Extract LicenseType, licenseTypeArray and licenseTypeLabels to own file
